requirements.txt
certifi==2022.9.24
numpy==1.23.5
pygrib==2.1.4
pyproj==3.4.0
python-dotenv==0.21.0
